Time: 0 ms (100%), Space: 7.9 MB (1.37%) - LeetHub

+++ 0371-sum-of-two-integers/0371-sum-of-two-integers.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int getSum(int a, int b) {
         while(b!=0){
-            int carry=(unsigned int) ( a & b );
+            int carry=( a & b );
             a=a^b;
             b = carry << 1;
         }
